% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineweight.R
\name{weights_spline}
\alias{weights_spline}
\alias{weights_spline,OneStageBasket-method}
\title{Weighting via Monotonic Splines}
\usage{
weights_spline(design, ...)

\S4method{weights_spline}{OneStageBasket}(
  design,
  n,
  diffknots = c(1, 0.5, 0),
  weightknots = c(0, 0, 1),
  splinemethod = "monoH.FC",
  clamplim = c(0, 1),
  ...
)
}
\arguments{
\item{design}{An object of class \code{Basket} created by
\code{setupOneStageBasket} or \code{setupTwoStageBasket}.}

\item{...}{Further arguments.}

\item{n}{The sample size per basket.}

\item{diffknots}{A vector of length k. Determines the levels of difference
at which the corresponding weights, specified in the vector
\code{weightknots}, are used.}

\item{weightknots}{A vector of length k. Determines the weights for sharing
at the corresponding level of difference specified in the vector
\code{diffknots}.}

\item{splinemethod}{The method used for interpolating between the knots
defined by \code{diffknots} and \code{weightknots}. Possible methods are
identical to the ones implemented in \code{stats::splinefun}, namely
"fmm", "periodic", "natural", "monoH.FC" and "hyman".}

\item{clamplim}{A vector of length two. Determines the lower bound and
upper bound used for clamping.}

\item{weights_spline}{Using splines for weighting whenever weighting is used
in baskexact, e.g. \code{toer}, \code{pow} or \code{ecd}.}
}
\value{
A matrix including the weights of all possible pairwise outcomes.
}
\description{
Weighting via Monotonic Splines
}
\details{
\code{weights_spline} calculates the weights based on monotonic
splines. First, an interpolating spline is defined by the user, identical
to the interpolating splines available in \code{\link{splinefun}}.
As such, more than two knots have to be specified via the arguments
\code{weightknots} and \code{diffknots}, which have to be vectors of
matching lengths. Values of corresponding positions within the vectors
form the knots used for interpolation.
The weight for two baskets i and j is then found by calculating the
difference in response rates for binary outcomes and applying the
interpolating spline.

The \code{clamplim} argument determined specific lower and upper bounds,
respectively. If a method for interpolation results in values outside these
bounds, they will be set to the given bound. The default is a lower bound
of 0 and an upper bound of 1.

The function is generally not called by the user but passed to another
function such as \code{\link{toer}} and \code{\link{pow}} to specify
how the weights are calculated.
}
\section{Methods (by class)}{
\itemize{
\item \code{weights_spline(OneStageBasket)}: Weights for a single-stage basket
design based on monotonic splines.

}}
\examples{
design <- baskexact::setupOneStageBasket(k = 3, p0 = 0.2)
baskexact::toer(design, n = 15, lambda = 0.99, weight_fun = weights_spline)

}
